requirements.txt
aniso8601==9.0.1
attrs==23.1.0
Babel==2.13.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
exceptiongroup==1.1.3
Flask==2.1.1
Flask-BabelEx==0.9.4
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Principal==0.4.0
flask-restplus==0.13.0
Flask-Security==3.0.0
Flask-WTF==1.2.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
pymongo==4.6.0
pytest==7.4.3
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.30.2
rpds-py==0.12.0
six==1.16.0
speaklater==1.3
tomli==2.0.1
Werkzeug==3.0.1
WTForms==3.1.1
